Factor out common code from DoFnFunction and MultiDoFnFunction.
